Move the different defaultOptions to Options.hs
Cyclic imports prevented these to be in the same
module as the declaration of the records. This
is eliminated so put the default options at the
record declaration.
