+++ src/Feldspar/Compiler.hs
@@ -44,16 +44,6 @@ module Feldspar.Compiler
   , programOptsArgs
   , getCore
   , printCore
-  , Options(..)
-  , defaultOptions
-  , sicsOptions
-  , sicsOptions2
-  , sicsOptions3
-  , FeldOpts(..)
-  , Target(..)
-  , c99PlatformOptions
-  , c99OpenMpPlatformOptions
-  , tic64xPlatformOptions
   -- * Internal functions
   , compileToCCore'
   , writeFiles
@@ -319,38 +309,6 @@ genIncludeLines opts mainHeader = concatMap include incs ++ "\n\n"
     include fname@('<':_) = "#include " ++ fname ++ "\n"
     include fname         = "#include \"" ++ fname ++ "\"\n"
     incs = includes (platform opts) ++ [fromMaybe "" mainHeader]
-
--- | Predefined options
-
-defaultOptions :: Options
-defaultOptions
-    = Options
-    { platform          = c99
-    , printHeader       = False
-    , useNativeArrays   = False
-    , useNativeReturns  = False
-    , frontendOpts      = defaultFeldOpts
-    , safetyLimit       = 2000
-    , nestSize          = 2
-    }
-
-c99PlatformOptions :: Options
-c99PlatformOptions              = defaultOptions
-
-c99OpenMpPlatformOptions :: Options
-c99OpenMpPlatformOptions        = defaultOptions { platform = c99OpenMp }
-
-tic64xPlatformOptions :: Options
-tic64xPlatformOptions           = defaultOptions { platform = tic64x }
-
-sicsOptions :: Options
-sicsOptions = defaultOptions { frontendOpts = defaultFeldOpts { targets = [SICS,CSE] }}
-
-sicsOptions2 :: Options
-sicsOptions2 = defaultOptions { frontendOpts = defaultFeldOpts { targets = [SICS] }}
-
-sicsOptions3 :: Options
-sicsOptions3 = defaultOptions { platform = c99Wool, frontendOpts = defaultFeldOpts { targets = [SICS,CSE,Wool] }}
 
 data BackendPass = BPFromCore
                  | BPArrayOps
